adjusting navigation button functionality
making /chapter/1 the abstract of the work.
adding a button "abstract" on frontpage.

As the abstract is not really interesting for some readers, it can be skipped. Therefore:
making /chapter/2 the actual first page of the text.
linking the "next" button to /chapter/2 and calling it "start the text" on frontpage and abstract page.

the "previous" button on /chapter/2 displays accordingly "abstract" instead of "previous"