fix battery percentage
due to fluctuating voltage when sunny, which does not directly relate to the charge, I lowered the battery threshold to 13,4V as it more accuratley displays the battery charge when not sunny. When sunny, it rises up to 14,6V, then it shows 100% charge by default as power is then supplied fully by the solar panel and no load is on the battery.